+++ app/ootd/search/page.tsx
@@ -1,0 +1,8 @@
+import WriteHeader from "../../../components/ootd/write/HeaderWrite";
+export default function OotdSearch() {
+  return (
+    <main>
+      <WriteHeader />
+    </main>
+  );
+}
